WIP: Update task execution engine and completion detector

+++ ifogsim/src/main/java/org/patch/processing/TaskCompletionDetector.java
@@ -212,13 +212,14 @@ public class TaskCompletionDetector {
             long energyConsumed, double cost) {
         try {
             // Create completion report
+            // Proto fields expect double, so convert long to double
             org.patch.proto.IfogsimScheduler.CompletedTask completedTask = org.patch.proto.IfogsimScheduler.CompletedTask
                     .newBuilder()
                     .setTaskId(taskId)
                     .setAssignedNodeId(String.valueOf(fogDevice.getId()))
-                    .setActualExecutionTimeMs(executionTime)
-                    .setActualLatencyMs(executionTime) // Using execution time as latency
-                    .setEnergyConsumed(energyConsumed)
+                    .setActualExecutionTimeMs((double) executionTime)
+                    .setActualLatencyMs((double) executionTime) // Using execution time as latency
+                    .setEnergyConsumed((double) energyConsumed)
                     .setDeadlineMet(success)
                     .setStartTime((long) CloudSim.clock())
                     .setCompletionTime((long) CloudSim.clock())
